Reverted previous commit and fixed/simplified the decimal parsing logic
- The decimal values are indeed parsed by reconstructing from three tokens, which can be two number tokens and either a slash or a dot, e.g. `1.23` or `1/3`.

+++ src/Pixeval/Filters/IQueryToken.cs
@@ -66,7 +66,7 @@ public interface IQueryToken
     }
 
     [DebuggerDisplay("{Value}")]
-    public readonly record struct Numeric(double Value) : INullableToken
+    public readonly record struct Numeric(long Value) : INullableToken
     {
         public bool IsNotEmpty() => Value >= 0;
